requirements.txt
langchain==0.3.27
langchain-community==0.3.29
langchain-openai==0.3.35
openai==2.8.1
pydantic==2.9.2
langfuse==3.10.1
python-dotenv==1.2.1
faiss-cpu==1.13.0